Split inference code out into a separate module
